Removing warnings in compiling. Most import change, I swaped use of m_point struct to CvPoint2D32f to remove warnings. I must determine if there is really necessity for 4 different classes/structs to represent 2D points: -- fftw_complex -- mcomplex and std::complex -- CvPoint2D32f -- mpoint
git-svn-id: file:///home/adenilson/temp/wtf/repository@114 56e0a439-041b-0410-8790-949b9ad245cc

+++ thesis/src/contour.cpp
@@ -90,7 +90,7 @@ void mark_centroid(CvSeq *contour, IplImage *img)
 		meanx /= contour->total;
 		meany /= contour->total;
 
-		cvRectangle(img, cvPoint(meanx-1, meany-1),
+		cvRectangle(img, cvPoint((int) meanx-1, (int) meany-1),
 			    cvPoint(meanx+1, meany+1),
 			    CV_RGB(0, 0, 255), 1);
